serde cleanup
