+++ crates/sui-types/src/sui_serde.rs
@@ -8,7 +8,6 @@ use std::marker::PhantomData;
 use std::ops::Deref;
 use std::str::FromStr;
 
-use fastcrypto::encoding::Hex;
 use move_core_types::account_address::AccountAddress;
 use move_core_types::language_storage::{StructTag, TypeTag};
 use schemars::JsonSchema;
@@ -28,7 +27,7 @@ use crate::{
 };
 
 #[inline]
-fn to_custom_error<'de, D, E>(e: E) -> D::Error
+pub(crate) fn to_custom_deser_error<'de, D, E>(e: E) -> D::Error
 where
     E: Debug,
     D: Deserializer<'de>,
@@ -37,7 +36,7 @@ where
 }
 
 #[inline]
-fn to_custom_ser_error<S, E>(e: E) -> S::Error
+pub(crate) fn to_custom_ser_error<S, E>(e: E) -> S::Error
 where
     E: Debug,
     S: Serializer,
@@ -95,61 +94,6 @@ where
         } else {
             R::deserialize_as(deserializer)
         }
-    }
-}
-
-/// custom serde for AccountAddress
-pub struct HexAccountAddress;
-
-impl SerializeAs<AccountAddress> for HexAccountAddress {
-    fn serialize_as<S>(value: &AccountAddress, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        Hex::serialize_as(value, serializer)
-    }
-}
-
-impl<'de> DeserializeAs<'de, AccountAddress> for HexAccountAddress {
-    fn deserialize_as<D>(deserializer: D) -> Result<AccountAddress, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let s = String::deserialize(deserializer)?;
-        if s.starts_with("0x") {
-            AccountAddress::from_hex_literal(&s)
-        } else {
-            AccountAddress::from_hex(&s)
-        }
-        .map_err(to_custom_error::<'de, D, _>)
-    }
-}
-
-/// Serializes a bitmap according to the roaring bitmap on-disk standard.
-/// <https://github.com/RoaringBitmap/RoaringFormatSpec>
-pub struct SuiBitmap;
-
-impl SerializeAs<roaring::RoaringBitmap> for SuiBitmap {
-    fn serialize_as<S>(source: &roaring::RoaringBitmap, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let mut bytes = vec![];
-
-        source
-            .serialize_into(&mut bytes)
-            .map_err(to_custom_ser_error::<S, _>)?;
-        Bytes::serialize_as(&bytes, serializer)
-    }
-}
-
-impl<'de> DeserializeAs<'de, roaring::RoaringBitmap> for SuiBitmap {
-    fn deserialize_as<D>(deserializer: D) -> Result<roaring::RoaringBitmap, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let bytes: Vec<u8> = Bytes::deserialize_as(deserializer)?;
-        roaring::RoaringBitmap::deserialize_from(&bytes[..]).map_err(to_custom_error::<'de, D, _>)
     }
 }
 
@@ -364,5 +308,42 @@ impl<'de> DeserializeAs<'de, ProtocolVersion> for AsProtocolVersion {
     {
         let b = BigInt::<u64>::deserialize(deserializer)?;
         Ok(ProtocolVersion::from(*b))
+    }
+}
+
+/// Serializes and deserializes a RoaringBitmap with its own on-disk standard.
+/// <https://github.com/RoaringBitmap/RoaringFormatSpec>
+pub(crate) struct SuiBitmap;
+
+impl SerializeAs<roaring::RoaringBitmap> for SuiBitmap {
+    fn serialize_as<S>(source: &roaring::RoaringBitmap, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        let mut bytes = vec![];
+
+        source
+            .serialize_into(&mut bytes)
+            .map_err(to_custom_ser_error::<S, _>)?;
+        Bytes::serialize_as(&bytes, serializer)
+    }
+}
+
+impl<'de> DeserializeAs<'de, roaring::RoaringBitmap> for SuiBitmap {
+    fn deserialize_as<D>(deserializer: D) -> Result<roaring::RoaringBitmap, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let bytes: Vec<u8> = Bytes::deserialize_as(deserializer)?;
+        let orig_bitmap = roaring::RoaringBitmap::deserialize_from(&bytes[..])
+            .map_err(to_custom_deser_error::<'de, D, _>)?;
+        let mut seen = std::collections::BTreeSet::new();
+        let mut new_bitmap = roaring::RoaringBitmap::new();
+        for v in orig_bitmap.iter() {
+            if seen.insert(v) {
+                new_bitmap.insert(v);
+            }
+        }
+        Ok(new_bitmap)
     }
 }
